support http header to indicate an expected failure
By setting an extra HTTP header 'X-Pacman-Expected-Failure' the server can
indicate that the failure is expected. The next server is tried without
error message and without increasing the server's error count.

This can be used by servers that are not expected to be complete, for
example when serving a local cache.

Signed-off-by: Christian Hesse <mail@eworm.de>