implement CacheServer
This implements a new configuration option 'CacheServer'. Adding a cache
server makes it ignore the server error limit.

We have a struct that stores the server errors. Extend (and rename) this
struct to store if this is a cache server. The errors are not increased
for cache servers, thus they are never ignored.

Signed-off-by: Christian Hesse <mail@eworm.de>